move namespace to svgcanvas folder
