+++ cypress/integration/unit/select.js
@@ -1,5 +1,5 @@
 import * as select from '../../../instrumented/svgcanvas/select.js';
-import { NS } from '../../../instrumented/common/namespaces.js';
+import { NS } from '../../../instrumented/svgcanvas/namespaces.js';
 
 describe('select', function () {
   const sandbox = document.createElement('div');
