+++ cypress/integration/unit/utilities.js
@@ -1,6 +1,6 @@
 /* eslint-disable max-len */
 import * as utilities from '../../../instrumented/svgcanvas/utilities.js';
-import { NS } from '../../../instrumented/common/namespaces.js';
+import { NS } from '../../../instrumented/svgcanvas/namespaces.js';
 
 describe('utilities', function () {
   /**
