+++ src/svgcanvas/draw.js
@@ -8,7 +8,7 @@
 import Layer from './layer.js';
 import HistoryRecordingService from './historyrecording.js';
 
-import { NS } from '../common/namespaces.js';
+import { NS } from './namespaces.js';
 import {
   toXml, getElem
 } from './utilities.js';
