+++ src/svgcanvas/json.js
@@ -6,7 +6,7 @@
  * @copyright 2010 Alexis Deveria, 2010 Jeff Schiller
  */
 import { getElem, assignAttributes, cleanupElement } from './utilities.js';
-import { NS } from '../common/namespaces.js';
+import { NS } from './namespaces.js';
 
 let jsonContext_ = null;
 let svgdoc_ = null;
