+++ src/svgcanvas/namespaces.js
