+++ src/svgcanvas/path-actions.js
@@ -6,7 +6,7 @@
  * @copyright 2011 Alexis Deveria, 2011 Jeff Schiller
  */
 
-import { NS } from '../common/namespaces.js';
+import { NS } from './namespaces.js';
 import { shortFloat } from '../common/units.js';
 import { ChangeElementCommand, BatchCommand } from './history.js';
 import {
