+++ src/svgcanvas/sanitize.js
@@ -6,7 +6,7 @@
  * @copyright 2010 Alexis Deveria, 2010 Jeff Schiller
  */
 
-import { getReverseNS, NS } from '../common/namespaces.js';
+import { getReverseNS, NS } from './namespaces.js';
 import { isGecko } from '../common/browser.js';
 import { getHref, setHref, getUrlFromAttr } from './utilities.js';
