+++ src/svgcanvas/selected-elem.js
@@ -6,7 +6,7 @@
  * @copyright 2010 Alexis Deveria, 2010 Jeff Schiller
  */
 
-import { NS } from '../common/namespaces.js';
+import { NS } from './namespaces.js';
 import * as hstry from './history.js';
 import * as pathModule from './path.js';
 import {
