+++ src/svgcanvas/svg-exec.js
@@ -22,7 +22,7 @@ import {
 } from '../common/units.js';
 import { isGecko, isChrome, isWebkit } from '../common/browser.js';
 import * as pathModule from './path.js';
-import { NS } from '../common/namespaces.js';
+import { NS } from './namespaces.js';
 import * as draw from './draw.js';
 import {
   recalculateDimensions
