+++ src/svgcanvas/svgroot.js
@@ -5,7 +5,7 @@
  *
  * @copyright 2010 Alexis Deveria, 2010 Jeff Schiller
  */
-import { NS } from '../common/namespaces.js';
+import { NS } from './namespaces.js';
 import { text2xml } from './utilities.js';
 
 /**
